+++ server/internal/model/entity/sys_config.go
@@ -24,6 +24,6 @@ type SysConfig struct {
 	AllowDel   int         `json:"allowDel"   orm:"allow_del"   description:"允许删除:0=否,1=是"`                                                                          // 允许删除:0=否,1=是
 	CreatedBy  uint64      `json:"createdBy"  orm:"created_by"  description:"创建人"`                                                                                   // 创建人
 	UpdatedBy  uint64      `json:"updatedBy"  orm:"updated_by"  description:"更新人"`                                                                                   // 更新人
-	CreateTime uint64      `json:"createTime" orm:"create_time" description:"创建时间"`                                                                                  // 创建时间
-	UpdateTime uint64      `json:"updateTime" orm:"update_time" description:"更新时间"`                                                                                  // 更新时间
+	CreateTime uint        `json:"createTime" orm:"create_time" description:"创建时间"`                                                                                  // 创建时间
+	UpdateTime uint        `json:"updateTime" orm:"update_time" description:"更新时间"`                                                                                  // 更新时间
 }
