+++ server/internal/model/entity/sys_cron_group.go
@@ -11,6 +11,6 @@ type SysCronGroup struct {
 	Sort      int    `json:"sort"      orm:"sort"       description:"排序（越小越靠前）"`  // 排序（越小越靠前）
 	Remark    string `json:"remark"    orm:"remark"     description:"备注"`         // 备注
 	Status    int    `json:"status"    orm:"status"     description:"状态:0禁用,1启用"` // 状态:0禁用,1启用
-	CreatedAt uint64 `json:"createdAt" orm:"created_at" description:"创建时间"`       // 创建时间
-	UpdatedAt uint64 `json:"updatedAt" orm:"updated_at" description:"更新时间"`       // 更新时间
+	CreatedAt uint   `json:"createdAt" orm:"created_at" description:"创建时间"`       // 创建时间
+	UpdatedAt uint   `json:"updatedAt" orm:"updated_at" description:"更新时间"`       // 更新时间
 }
